refactor: extract comments_controller concerns
Extract approval actions, conversion, and batch operations from comments_controller into separate concern modules. This reduces the controller from 493 to 314 lines and improves code organization.

- Add ApprovalActions concern (approve, update_action)
- Add Conversion concern (convert and helper methods)
- Add BatchOperations concern (batch_destroy, move)
