+++ engines/collavre/app/controllers/collavre/comments_controller.rb
@@ -1,5 +1,9 @@
 module Collavre
   class CommentsController < ApplicationController
+    include Collavre::Comments::ApprovalActions
+    include Collavre::Comments::Conversion
+    include Collavre::Comments::BatchOperations
+
     before_action :set_creative
     before_action :set_comment, only: [ :destroy, :show, :update, :convert, :approve, :update_action ]
 
@@ -248,132 +252,7 @@ module Collavre
       end
     end
 
-    def convert
-      unless can_convert_comment?
-        render json: { error: I18n.t("collavre.comments.convert_not_allowed") }, status: :forbidden and return
-      end
 
-      created_creatives = ::MarkdownImporter.import(
-        @comment.content,
-        parent: @creative,
-        user: @creative.user,
-        create_root: true
-      )
-
-      primary_creative = created_creatives.first
-      system_message = build_convert_system_message(primary_creative) if primary_creative
-
-      @comment.destroy
-
-      if system_message.present?
-        Current.set(session: nil) do
-          @creative.comments.create!(content: system_message, user: nil)
-        end
-      end
-
-      head :no_content
-    end
-
-    def approve
-      status = @comment.approval_status(Current.user)
-      if status != :ok
-        error_key = case status
-        when :invalid_action_format then "collavre.comments.approve_invalid_format"
-        when :missing_action then "collavre.comments.approve_missing_action"
-        when :missing_approver then "collavre.comments.approve_missing_approver"
-        when :admin_required then "collavre.comments.approve_admin_required"
-        else "collavre.comments.approve_not_allowed"
-        end
-        http_status = case status
-        when :invalid_action_format, :missing_action, :missing_approver
-                        :unprocessable_entity
-        else
-                        :forbidden
-        end
-        render json: { error: I18n.t(error_key) }, status: http_status and return
-      end
-
-      begin
-        ::Comments::ActionExecutor.new(comment: @comment, executor: Current.user).call
-        @comment = Comment.with_attached_images.includes(:comment_reactions, :comment_versions, :selected_version).find(@comment.id)
-        render partial: "collavre/comments/comment", locals: { comment: @comment, current_topic_id: current_topic_context }
-      rescue ::Comments::ActionExecutor::ExecutionError => e
-        render json: { error: e.message }, status: :unprocessable_entity
-      end
-    end
-
-    def update_action
-      # Initial checks outside the lock
-      executed_error = false
-      update_success = false
-      approver_mismatch_error = false
-      status_error_key = nil
-      status_http_status = nil
-      validation_error_message = nil
-
-      @comment.with_lock do
-        @comment.reload
-
-        status_in_lock = @comment.approval_status(Current.user)
-        # Allow repairing invalid format if user is approver
-        if status_in_lock == :invalid_action_format
-          if @comment.approver_id == Current.user&.id
-            status_in_lock = :ok
-          else
-            status_in_lock = :not_allowed
-          end
-        end
-
-        if status_in_lock != :ok
-          approver_mismatch_error = true
-          status_error_key = case status_in_lock
-          when :invalid_action_format then "collavre.comments.approve_invalid_format"
-          when :missing_action then "collavre.comments.approve_missing_action"
-          when :missing_approver then "collavre.comments.approve_missing_approver"
-          when :admin_required then "collavre.comments.approve_admin_required"
-          else "collavre.comments.approve_not_allowed"
-          end
-          status_http_status = case status_in_lock
-          when :invalid_action_format, :missing_action, :missing_approver
-                          :unprocessable_entity
-          else
-                          :forbidden
-          end
-        elsif @comment.action_executed_at.present?
-          executed_error = true
-        else
-          action_payload = params.dig(:comment, :action)
-          if action_payload.blank?
-            validation_error_message = I18n.t("collavre.comments.approve_missing_action")
-          else
-            begin
-              validator = ::Comments::ActionValidator.new(comment: @comment)
-              parsed_payload = validator.validate!(action_payload)
-              normalized_action = JSON.pretty_generate(parsed_payload)
-              update_success = @comment.update(action: normalized_action)
-            rescue ::Comments::ActionValidator::ValidationError => e
-              validation_error_message = e.message
-            end
-          end
-        end
-      end
-
-      if approver_mismatch_error
-        render json: { error: I18n.t(status_error_key) }, status: status_http_status
-      elsif validation_error_message
-        render json: { error: validation_error_message }, status: :unprocessable_entity
-      elsif executed_error
-        render json: { error: I18n.t("collavre.comments.approve_already_executed") }, status: :unprocessable_entity
-      elsif update_success
-        @comment = Comment.with_attached_images.includes(:comment_reactions, :comment_versions, :selected_version).find(@comment.id)
-        render partial: "collavre/comments/comment", locals: { comment: @comment, current_topic_id: current_topic_context }
-      else
-        error_message = @comment.errors.full_messages.to_sentence.presence || I18n.t("collavre.comments.action_update_error")
-        render json: { error: error_message }, status: :unprocessable_entity
-      end
-    rescue ::Comments::ActionValidator::ValidationError => e
-      render json: { error: e.message }, status: :unprocessable_entity
-    end
 
     def show
       redirect_to creative_path(@creative, comment_id: @comment.id)
@@ -406,53 +285,6 @@ module Collavre
       render json: CommandMenuService.new(user: Current.user).items
     end
 
-    MAX_BATCH_DELETE = 100
-
-    def batch_destroy
-      comment_ids = Array(params[:comment_ids]).map(&:to_i).uniq.first(MAX_BATCH_DELETE)
-      if comment_ids.empty?
-        render json: { error: I18n.t("collavre.comments.batch_delete_no_selection") }, status: :unprocessable_entity and return
-      end
-
-      is_admin = @creative.has_permission?(Current.user, :admin)
-      is_creative_owner = @creative.user == Current.user
-
-      visible_scope = @creative.comments.where(
-        "comments.private = ? OR comments.user_id = ? OR comments.approver_id = ?",
-        false, Current.user.id, Current.user.id
-      )
-      comments = visible_scope.where(id: comment_ids).to_a
-
-      if comments.length != comment_ids.length
-        render json: { error: I18n.t("collavre.comments.batch_delete_not_found") }, status: :not_found and return
-      end
-
-      # Check permissions: user must own all comments, or be admin/creative owner
-      unless is_admin || is_creative_owner
-        unauthorized = comments.reject { |c| c.user == Current.user }
-        if unauthorized.any?
-          render json: { error: I18n.t("collavre.comments.not_owner") }, status: :forbidden and return
-        end
-      end
-
-      Comment.where(id: comments.map(&:id)).destroy_all
-
-      head :no_content
-    end
-
-    def move
-      result = CommentMoveService.new(creative: @creative, user: Current.user).call(
-        comment_ids: params[:comment_ids],
-        target_creative_id: params[:target_creative_id],
-        target_topic_id: params[:target_topic_id]
-      )
-      render json: result
-    rescue CommentMoveService::MoveError => e
-      status = e.message == I18n.t("collavre.comments.move_not_allowed") ? :forbidden : :unprocessable_entity
-      render json: { error: e.message }, status: status
-    rescue ActiveRecord::RecordInvalid => e
-      render json: { error: e.record.errors.full_messages.to_sentence.presence || I18n.t("collavre.comments.move_error") }, status: :unprocessable_entity
-    end
 
     private
 
@@ -473,17 +305,6 @@ module Collavre
 
     def comment_params
       params.require(:comment).permit(:content, :private, :topic_id, :quoted_comment_id, :quoted_text, :review_type, images: [])
-    end
-
-    def can_convert_comment?
-      @comment.user == Current.user || @creative.has_permission?(Current.user, :admin)
-    end
-
-    def build_convert_system_message(creative)
-      title = helpers.strip_tags(creative.description).to_s.strip
-      title = I18n.t("collavre.comments.convert_system_message_default_title") if title.blank?
-      url = creative_path(creative)
-      I18n.t("collavre.comments.convert_system_message", title: title, url: url)
     end
 
     def current_topic_context
